remove hint about tcpflags proto filter
Best not explicitly name that functionality, as it does not give the
expected results when the statement containing the tcpflags match is
negated later on. Even if tcpflags ensures proto tcp, matching for the
absence of a specific tcp flags retunrs UDP of course, as the negation
occures at a later stage based on the truth value of the tcpflags
matcher.